refactor: rename context to contexter
