+++ src/scripter.ts
@@ -1,4 +1,4 @@
-import type { Context } from './context';
+import type { Context } from './contexter';
 import { run } from './runner';
 
 interface Scripter {
